Update installation_chair.md
Discussing the role of an installation chair with one of the NIME20 Installation Co-Chairs it was brought to our attention that installations are not necessarily fun. They can be engaging but at the same time deal with serious and or reflective topics that by definition cannot be considered fun. For this reason it was suggested to change the assessment criteria from "fun factor" to "engagement factor", to reflect better what the NIME chairs are looking for in installation proposals.